UI tweaks, moved "Players Using the Mod" to the HUD renderer so you can always show it with a toggle.
