+++ src/main/resources/assets/wildfire_gender/lang/en_us.json
@@ -11,7 +11,7 @@
 	"wildfire_gender.player_list.bounce_multiplier": "Bounce Multiplier: %sx",
 	"wildfire_gender.player_list.breast_momentum": "Breast Momentum: %s%%",
 	"wildfire_gender.player_list.female_sounds": "Female Sounds: %s",
-	"wildfire_gender.wardrobe.players_using_mod": "Players using the mod:",
+	"wildfire_gender.wardrobe.players_using_mod": "Players Using the Mod:",
 	"wildfire_gender.always_show_list": "Always Show List: %s",
 
 	"wildfire_gender.wardrobe.title": "Wildfire's Female Gender Mod",
